Eliminación de rutinas duplicadas ya existentes en base.py, otras correcciones del estilo, excepción clara si el método de pago no es válido o el pago no existe, validación de pago con tarjeta corregida
